requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
email-validator
python-jose[cryptography]
passlib[bcrypt]
google-auth
requests
sqlalchemy
alembic
python-dotenv
